mend

+++ src/DependencyInjection/Compiler/ApplicationCompilerPass.php
@@ -10,7 +10,6 @@ use Northrook\Resource\Path;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\Yaml\Yaml;
-use function Assert\isCLI;
 use UnexpectedValueException;
 use Override;
 
@@ -23,7 +22,7 @@ final readonly class ApplicationCompilerPass implements CompilerPassInterface
     {
         $this->projectDirectory = $container->getParameter( 'kernel.project_dir' );
 
-        Output::info( "Using project directory: " . $this->projectDirectory );
+        Output::info( 'Using project directory: '.$this->projectDirectory );
 
         $this->appKernel()
             ->publicIndex()
@@ -35,18 +34,18 @@ final readonly class ApplicationCompilerPass implements CompilerPassInterface
 
     public function appKernel() : self
     {
-        $this->createFile(
+        $this->createPhpFile(
             'src/Kernel.php',
             <<<PHP
                 <?php
-                    
+                
                 declare(strict_types=1);
-                    
+                
                 namespace App;
-                    
+                
                 use Symfony\Bundle\FrameworkBundle\Kernel as FrameworkKernel;
                 use Symfony\Component\HttpKernel\Kernel as HttpKernel;
-                    
+                
                 final class Kernel extends HttpKernel
                 {
                     use FrameworkKernel\MicroKernelTrait;
@@ -58,15 +57,15 @@ final readonly class ApplicationCompilerPass implements CompilerPassInterface
 
     public function publicIndex() : self
     {
-        $this->createFile(
+        $this->createPhpFile(
             'public/index.php',
             <<<PHP
                 <?php
-                    
+                
                 declare(strict_types=1);
-                    
+                
                 require_once \dirname(__DIR__).'/vendor/autoload_runtime.php';
-                    
+                
                 return function (array \$context): App\Kernel {
                     return new App\Kernel(\$context['APP_ENV'], (bool) \$context['APP_DEBUG']);
                 };
@@ -104,16 +103,20 @@ final readonly class ApplicationCompilerPass implements CompilerPassInterface
 
     public function appControllerRouteConfiguration() : self
     {
+        if ( $this->path( 'routes.php' )->exists ) {
+            return $this;
+        }
+
         $this->removeFile( 'config/routes.yaml' );
-        $this->createFile(
+        $this->createPhpFile(
             'routes.php',
             <<<PHP
                 <?php
-                    
+                
                 declare( strict_types = 1 );
-                    
+                
                 use Symfony\Component\Routing\Loader\Configurator\RoutingConfigurator;
-                    
+                
                 return static function ( RoutingConfigurator \$routes ) : void {
                     \$routes->import(
                         [
@@ -131,26 +134,29 @@ final readonly class ApplicationCompilerPass implements CompilerPassInterface
 
     public function createConfigServices() : self
     {
+        if ( $this->path( 'config/services.php' )->exists ) {
+            return $this;
+        }
         $this->removeFile( 'config/services.yaml' );
-        $this->createFile(
+        $this->createPhpFile(
             'config/services.php',
             <<<PHP
                 <?php
-                    
+                
                 declare( strict_types = 1 );
-                    
+                
                 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-                    
+                
                 return static function ( ContainerConfigurator \$container ) : void {
-                    
+                
                     \$services = \$container->services();
-                    
+                
                     // Defaults for App services.
                     \$services
                         ->defaults()
                         ->autowire()
                         ->autoconfigure();
-                    
+                
                     \$services
                         // Make classes in src/ available to be used as services.
                         ->load( "App\\\\", __DIR__ . '/../src/' )
@@ -170,13 +176,13 @@ final readonly class ApplicationCompilerPass implements CompilerPassInterface
 
     public function configurePreload() : self
     {
-        $this->createFile(
+        $this->createPhpFile(
             'config/preload.php',
             <<<'PHP'
                 <?php
-                    
+                
                 declare(strict_types=1);
-                    
+                
                 if (\file_exists(\dirname(__DIR__).'/var/cache/prod/App_KernelProdContainer.preload.php')) {
                     \opcache_compile_file(\dirname(__DIR__).'/var/cache/prod/App_KernelProdContainer.preload.php');
                 }
@@ -193,38 +199,30 @@ final readonly class ApplicationCompilerPass implements CompilerPassInterface
         return $this;
     }
 
-    private function removeFile( string $fromProjectDir ) : void
-    {
-        $path = new Path( "{$this->projectDirectory}/{$fromProjectDir}" );
-
-        if ( $path->delete() ) {
-            Output::info( "Compiler removed {$fromProjectDir}." );
-        }
-    }
-
     private function createYamlFile(
         string $fromProjectDir,
         #[Language( 'PHP' )] mixed  $input,
         bool   $overwrite = false,
     ) : void {
-        $path    = new Path( "{$this->projectDirectory}/{$fromProjectDir}" );
-        $content = Yaml::dump( $input );
+        $path = new Path( "{$this->projectDirectory}/{$fromProjectDir}" );
 
         if ( $path->exists && false === $overwrite ) {
             return;
         }
+
+        $content = Yaml::dump( $input );
 
         if ( $content && $path->save( $content ) ) {
             Output::info( "Compiler generated {$fromProjectDir}." );
         }
     }
 
-    private function createFile(
+    private function createPhpFile(
         string $fromProjectDir,
         #[Language( 'PHP' )] string $php,
         bool   $overwrite = false,
     ) : void {
-        $path = new Path( "{$this->projectDirectory}/{$fromProjectDir}" );
+        $path = $this->path( $fromProjectDir );
 
         if ( $path->exists && false === $overwrite ) {
             return;
@@ -254,5 +252,19 @@ final readonly class ApplicationCompilerPass implements CompilerPassInterface
             throw new UnexpectedValueException( 'Autoconfigure encountered an unexpected error preparing the PHP string.' );
         }
         return $content;
+    }
+
+    private function removeFile( string $fromProjectDir ) : void
+    {
+        $path = new Path( "{$this->projectDirectory}/{$fromProjectDir}" );
+
+        if ( $path->delete() ) {
+            Output::info( "Compiler removed {$fromProjectDir}." );
+        }
+    }
+
+    private function path( string $fromProjectDir ) : Path
+    {
+        return new Path( "{$this->projectDirectory}/{$fromProjectDir}" );
     }
 }
